+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/configuration/deserializer/SequentialInboxConfigurationUtils.java
@@ -5,11 +5,15 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 class SequentialInboxConfigurationUtils {
 
-    <T> T newInstance(String className) {
+    <T> T newInstance(String className, Class<?> expectedSupertype) {
         try {
             @SuppressWarnings("unchecked")
-            Class<T> conditionClass = (Class<T>) Class.forName(className);
-            return conditionClass.getDeclaredConstructor().newInstance();
+            Class<T> clazz = (Class<T>) Class.forName(className);
+            if (!expectedSupertype.isAssignableFrom(clazz)) {
+                throw SequentialInboxConfigurationException.badInstanceType(clazz, expectedSupertype);
+            }
+
+            return clazz.getDeclaredConstructor().newInstance();
         } catch (ReflectiveOperationException e) {
             throw SequentialInboxConfigurationException.errorWhileCreatingInstance(className, e);
         }