+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/configuration/model/SequencedMessageType.java
@@ -1,29 +1,50 @@
 package ch.admin.bit.jeap.messaging.sequentialinbox.configuration.model;
 
 import ch.admin.bit.jeap.messaging.avro.AvroMessage;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.util.Set;
 
-@Getter
 @AllArgsConstructor // for builder
 @NoArgsConstructor // for jackson
 @Builder
 public class SequencedMessageType {
 
     private String type;
+    @Getter(AccessLevel.PACKAGE)
+    private String subType;
+    @Getter
     private String topic;
+    @Getter
     private String clusterName;
+    @Getter
     private ContextIdExtractor<AvroMessage> contextIdExtractor;
+    @Getter
     private MessageFilter<AvroMessage> messageFilter;
+    @Getter
     private ReleaseCondition releaseCondition;
+
+    /**
+     * @return The message type name, postfixed with .&lt;subType&gt; if a subtype is configured
+     */
+    public String getQualifiedName() {
+        if (subType == null) {
+            return type;
+        }
+        return type + "." + subType;
+    }
+
+    /**
+     * @return The jEAP message type name for this sequenced message type, without any subtype postfix
+     */
+    public String getJeapMessageTypeName() {
+        return type;
+    }
 
     @Override
     public String toString() {
-        return "SequencedMessageType{type='%s',topic=%s,clusterName=%s,contextIdExtractor=%s,messageFilter=%s,releaseCondition=%s}".formatted(type,
+        return "SequencedMessageType{type='%s',subType='%s',topic=%s,clusterName=%s,contextIdExtractor=%s,messageFilter=%s,releaseCondition=%s}".formatted(type,
+                subType,
                 topic,
                 clusterName,
                 contextIdExtractor.getClass().getName(),