+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.30.0] - 2025-03-20
+
+### Added
+
+- The sequential inbox configuration now supports defining subtypes
+  - This allows to define a sequence for different subtypes of the same avro message type
+
 ## [8.29.0] - 2025-03-18
 
 ### Changed